MM-67920 unify e2e team settings tests (#35867)
* MM-67920 - extract duplicated policy editor helpers

* split team settings mega file into focused specs

* remove duplicate team icon test file

* rename Access Control to Membership Policies in e2e

* replace networkidle with explicit element waits

* fix linter

* fix attribute loading issue

* remove unnecessary comments

* fix linter
